Set up database and dotenv, created addUser function
Did the boilerplate to set up Postgres and created a dotenv file. Also
created a function to add the user to the database. As I'm looking over
it to write this I realize I should probably move where it hashes the
password and maybe tighten up the conditional logic, but for now it works
by checking to make sure the username or email address isn't already in
the database, if so it returns an error message otherwise it inserts the
username, password and hashed password into the user_info table and sends
a success response.