Set up the basics of JWT and connected a response to front end
Set up JWT so now when a user logs in they will receive a token in the
response. Set up the front end to display error messages received in
response to invalid inputs in the login and sign up fields. Still need to
implement navigation guards and token checks on the front end.