Websockets and bi-directional communication work
Set up Socket.io and users can now send and receive messages. The socket.io
client initializes in the chat container compolnent when the user reaches
the chat view. Also made some changes to the getUser backend function,
now along with a JWT token the username is sent in the response JSON and
then stored in session storage client-side. This is so the chat component
can pull that to render the username.